recsys_seq/requirements.txt
numpy
pandas
scipy
scikit-learn
tqdm
pyyaml
torch
